dbus/dmabuf_gl.py
"""
OpenGL DMA-BUF Renderer - Simplified approach

mmapでデータを読み取り、OpenGLテクスチャにアップロードして正規化
"""

import logging
import numpy as np
from OpenGL import EGL
from OpenGL.EGL import *
from OpenGL.GL import *
from ctypes import *

logger = logging.getLogger(__name__)


class DMABUFRenderer:
    """
    mmapデータをOpenGLで正規化してレンダリング
    """
    
    def __init__(self):
        self.egl_display = None
        self.egl_context = None
        self.egl_surface = None
        self.fbo = None
        self.texture = None
        self.initialized = False
        
    def initialize(self):
        """EGL初期化（ヘッドレス）"""
        try:
            logger.info("Initializing EGL for headless rendering...")
            
            # EGLディスプレイ取得
            self.egl_display = eglGetDisplay(EGL_DEFAULT_DISPLAY)
            if self.egl_display == EGL_NO_DISPLAY:
                raise RuntimeError("Failed to get EGL display")
            
            # EGL初期化
            major = c_int32()
            minor = c_int32()
            if not eglInitialize(self.egl_display, byref(major), byref(minor)):
                raise RuntimeError("Failed to initialize EGL")
            
            logger.info(f"✓ EGL initialized: {major.value}.{minor.value}")
            
            self.initialized = True
            return True
            
        except Exception as e:
            logger.error(f"EGL initialization failed: {e}")
            import traceback
            logger.error(traceback.format_exc())
            return False
    
    def render_from_mmap(self, data, width, height, stride, fourcc):
        """
        mmapデータをOpenGLテクスチャにアップロードして正規化
        
        Args:
            data: mmapデータ
            width, height: サイズ
            stride: ストライド
            fourcc: フォーマット
            
        Returns:
            RGB NumPy配列
        """
        try:
            # TODO: OpenGL実装
            # 現時点では非実装
            logger.warning("OpenGL rendering not implemented yet, falling back to CPU")
            return None
            
        except Exception as e:
            logger.error(f"OpenGL rendering failed: {e}")
            return None
    
    def cleanup(self):
        """クリーンアップ"""
        try:
            if self.egl_display:
                eglTerminate(self.egl_display)
                self.egl_display = None
            self.initialized = False
        except Exception as e:
            logger.error(f"Cleanup failed: {e}")
