requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
attrs==25.4.0
av==16.0.1
cffi==2.0.0
cryptography==46.0.3
dasbus==1.7
dnspython==2.8.0
frozenlist==1.8.0
google-crc32c==1.8.0
idna==3.11
ifaddr==0.2.0
multidict==6.7.0
numpy==2.4.1
pillow==12.1.0
propcache==0.4.1
pycairo==1.29.0
pycparser==2.23
pyee==13.0.0
PyGObject==3.54.5
pylibsrtp==1.0.0
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
pyOpenSSL==25.3.0
typing_extensions==4.15.0
yarl==1.22.0
